+++ Android/app/src/main/res/values/styles.xml
@@ -13,4 +13,5 @@
         <item name="windowNoTitle">true</item>
     </style>
 
+
 </resources>
